+++ CHANGELOG.md
@@ -109,6 +109,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   xcconfig files are missing.  
   [Samuel Giddins](https://github.com/segiddins)
 
+* Correctly detect that a prebuilt pod uses Swift.  
+  [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
+  [#8649](https://github.com/CocoaPods/CocoaPods/issues/8649)
+
 
 ## 1.7.5 (2019-07-19)